Moved components into separate files: GameListCards and GameListContainer
